@@ -6,19 +6,25 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "formik": "^2.4.6",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-helmet": "^6.1.0",
+    "react-router-dom": "^6.28.0",
     "react-scripts": "^5.0.1",
+    "react-transition-group": "^4.4.5",
+    "sass": "^1.38.0",
     "web-vitals": "^1.0.1",
     "webpack": "^5.91.0",
-    "sass": "^1.38.0"
+    "yup": "^1.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "dev": "vite"
   },
   "eslintConfig": {
     "extends": [
@@ -39,6 +45,9 @@
     ]
   },
   "devDependencies": {
-    "sass": "^1.38.0"
+    "@types/react": "^18.3.3",
+    "@vitejs/plugin-react": "^4.3.3",
+    "sass": "^1.38.0",
+    "vite": "^5.4.11"
   }
 }
