@@ -1,44 +1,44 @@
-import { Component } from "react/cjs/react.production.min";
+import { lazy, Suspense } from 'react';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import decoration from '../../resources/img/vision.png';
-import ErrorBoundary from '../errorBoundary/ErrorBoundary';
+import Spinner from '../spinner/Spinner';
+import Page404 from '../pages/404';  // Импортируем компонент напрямую
+import MainPage from '../pages/MainPage';  // Импортируем компонент напрямую
+import ComicsPage from '../pages/ComicsPage';  // Импортируем компонент напрямую
+import SingleComicLayout from '../pages/singleComicLayout/SingleComicLayout';  // Импортируем компонент напрямую
+import SingleCharacterLayout from '../pages/singleCharacterLayout/SingleCharacterLayout';  // Импортируем компонент напрямую
+import SinglePage from '../pages/SinglePage';  // Импортируем компонент напрямую
 
-class App extends Component{
-    state = {
-        selectedChar: null
-    }
-
-    onCharSelected = (id) => {
-        this.setState({
-            selectedChar: id
-        })
-    }
-
-    render() {
-        return (
+const App = () => {
+    return (
+        <Router>
             <div className="app">
-                <ErrorBoundary>
-                    <AppHeader/>
-                </ErrorBoundary>
+                <AppHeader/>
                 <main>
-                    <RandomChar/>
-                    <div className="char__content">
-                        <ErrorBoundary>
-                            <CharList onCharSelected={this.onCharSelected}/>
-                        </ErrorBoundary>
-                        <ErrorBoundary>
-                            <CharInfo charId={this.state.selectedChar}/>
-                        </ErrorBoundary>
-
-                    </div>
-                    <img className="bg-decoration" src={decoration} alt="vision"/>
+                    <Suspense fallback={<Spinner/>}>
+                        <Switch>
+                            <Route exact path="/">
+                                <MainPage/>
+                            </Route>
+                            <Route exact path="/comics">
+                                <ComicsPage/>
+                            </Route>
+                            <Route exact path="/comics/:id">
+                                <SinglePage Component={SingleComicLayout} dataType='comic'/>
+                            </Route>
+                            <Route exact path="/characters/:id">
+                                <SinglePage Component={SingleCharacterLayout} dataType='character'/>
+                            </Route>
+                            <Route path="*">
+                                <Page404/>
+                            </Route>
+                        </Switch>
+                    </Suspense>
                 </main>
             </div>
-        )
-    }
+        </Router>
+    )
 }
 
 export default App;
