@@ -1,18 +1,24 @@
+import {Link, NavLink} from 'react-router-dom';
 import './appHeader.scss';
 
 const AppHeader = () => {
     return (
         <header className="app__header">
             <h1 className="app__title">
-                <a href="https://www.marvel.com/">
+                <Link to="/">
                     <span>Marvel</span> information portal
-                </a>
+                </Link>
             </h1>
             <nav className="app__menu">
                 <ul>
-                    <li><a href="https://marvel.com/characters">Characters</a></li>
-                    /
-                    <li><a href="https://marvel.com/comics">Comics</a></li>
+                    <li><NavLink 
+                        end
+                        style={({isActive}) => ({color: isActive ? '#9f0013' : 'inherit'})}
+                        to="/">Characters</NavLink></li>
+                        /
+                    <li><NavLink 
+                        style={({isActive}) => ({color: isActive ? '#9f0013' : 'inherit'})}
+                        to="/comics">Comics</NavLink></li>
                 </ul>
             </nav>
         </header>
