@@ -1,88 +1,57 @@
-import React, { useState, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import MarvelService from '../services/MarvelService';
-import Spinner from '../spinner/Spinner';
-import ErrorMessage from '../errorMessage/ErrorMessage';
-import Skeleton from '../skeleton/Skeleton';
+
+import useMarvelService from '../../services/MarvelService';
+import setContent from '../utils/setContent';
+
 import './charInfo.scss';
 
-const CharInfo = ({ charId }) => {
-    // Состояния компонента
+const CharInfo = (props) => {
+
     const [char, setChar] = useState(null);
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState(false);
 
-    // Создание экземпляра MarvelService
-    const marvelService = useRef(new MarvelService());
+    const {getCharacter, clearError, process, setProcess} = useMarvelService();
 
-    // Вызываем updateChar при монтировании компонента и при изменении charId
     useEffect(() => {
-        updateChar();
-    }, [charId]);
+        updateChar()
+        // eslint-disable-next-line
+    }, [props.charId])
 
-    // Обновление информации о персонаже
     const updateChar = () => {
+        const {charId} = props;
         if (!charId) {
             return;
         }
 
-        // Устанавливаем состояние загрузки
-        onCharLoading();
-
-        // Получаем информацию о персонаже из сервиса Marvel
-        marvelService.current
-            .getCharacter(charId)
+        clearError();
+        getCharacter(charId)
             .then(onCharLoaded)
-            .catch(onError);
+            .then(() => setProcess('confirmed'))
     }
 
-    // Обработка успешной загрузки информации о персонаже
     const onCharLoaded = (char) => {
         setChar(char);
-        setLoading(false);
     }
-
-    // Установка состояния загрузки
-    const onCharLoading = () => {
-        setLoading(true);
-    }
-
-    // Обработка ошибки загрузки информации о персонаже
-    const onError = () => {
-        setLoading(false);
-        setError(true);
-    }
-
-    // Переменные для управления отображением контента
-    const skeleton = char || loading || error ? null : <Skeleton />;
-    const errorMessage = error ? <ErrorMessage /> : null;
-    const spinner = loading ? <Spinner /> : null;
-    const content = !(loading || error || !char) ? <View char={char} /> : null;
 
     return (
         <div className="char__info">
-            {skeleton}
-            {errorMessage}
-            {spinner}
-            {content}
+            {setContent(process, View, char)}
         </div>
-    );
+    )
 }
 
-// Компонент для отображения информации о персонаже
-const View = ({ char }) => {
-    const { name, description, thumbnail, homepage, wiki, comics } = char;
+const View = ({data}) => {
+    const {name, description, thumbnail, homepage, wiki, comics} = data;
 
-    // Установка стиля изображения
-    let imgStyle = { 'objectFit': 'cover' };
+    let imgStyle = {'objectFit' : 'cover'};
     if (thumbnail === 'http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') {
-        imgStyle = { 'objectFit': 'contain' };
+        imgStyle = {'objectFit' : 'contain'};
     }
 
     return (
         <>
             <div className="char__basics">
-                <img src={thumbnail} alt={name} style={imgStyle} />
+                <img src={thumbnail} alt={name} style={imgStyle}/>
                 <div>
                     <div className="char__info-name">{name}</div>
                     <div className="char__btns">
@@ -103,17 +72,18 @@ const View = ({ char }) => {
                 {comics.length > 0 ? null : 'There is no comics with this character'}
                 {
                     comics.map((item, i) => {
-                        if (i > 9) return null;
+                        // eslint-disable-next-line
+                        if (i > 9) return;
                         return (
                             <li key={i} className="char__comics-item">
                                 {item.name}
                             </li>
-                        );
+                        )
                     })
-                }
+                }                
             </ul>
         </>
-    );
+    )
 }
 
 CharInfo.propTypes = {
