@@ -1,71 +1,43 @@
-import React, { useState, useEffect, useCallback } from 'react';
-import Spinner from '../spinner/Spinner';
-import ErrorMessage from '../errorMessage/ErrorMessage';
-import MarvelService from '../services/MarvelService';
+import {useState, useEffect} from 'react';
+import useMarvelService from '../../services/MarvelService';
+import setContent from '../utils/setContent';
 
 import './randomChar.scss';
 import mjolnir from '../../resources/img/mjolnir.png';
 
 const RandomChar = () => {
-    // Состояния компонента
-    const [char, setChar] = useState({});
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(false);
 
-    // Экземпляр MarvelService
-    const marvelService = new MarvelService();
+    const [char, setChar] = useState(null);
+    const {getCharacter, clearError, process, setProcess} = useMarvelService();
 
-    // Эффект для загрузки персонажа при монтировании компонента
     useEffect(() => {
         updateChar();
-        // const timerId = setInterval(updateChar, 15000);
+        const timerId = setInterval(updateChar, 60000);
 
-        // Очистка интервала при размонтировании компонента
         return () => {
-            // clearInterval(timerId);
+            clearInterval(timerId)
         }
-    }, []);
+        // eslint-disable-next-line
+    }, [])
 
-    // Функция обработки успешной загрузки персонажа
     const onCharLoaded = (char) => {
         setChar(char);
-        setLoading(false);
     }
 
-    // Функция установки состояния загрузки
-    const onCharLoading = () => {
-        setLoading(true);
-    }
-
-    // Функция обработки ошибки
-    const onError = () => {
-        setLoading(false);
-        setError(true);
-    }
-
-    // Функция для обновления персонажа
-    const updateChar = useCallback(() => {
-        const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
-        onCharLoading();
-        marvelService
-            .getCharacter(id)
+    const updateChar = () => {
+        clearError();
+        const id = Math.floor(Math.random() * (1011400 - 1011000)) + 1011000;
+        getCharacter(id)
             .then(onCharLoaded)
-            .catch(onError);
-    }, []);
-
-    // Состояния для отображения контента
-    const errorMessage = error ? <ErrorMessage /> : null;
-    const spinner = loading ? <Spinner /> : null;
-    const content = !(loading || error) ? <View char={char} /> : null;
+            .then(() => setProcess('confirmed'));
+    }
 
     return (
         <div className="randomchar">
-            {errorMessage}
-            {spinner}
-            {content}
+            {setContent(process, View, char)}
             <div className="randomchar__static">
                 <p className="randomchar__title">
-                    Random character for today!<br />
+                    Random character for today!<br/>
                     Do you want to get to know him better?
                 </p>
                 <p className="randomchar__title">
@@ -74,23 +46,22 @@ const RandomChar = () => {
                 <button onClick={updateChar} className="button button__main">
                     <div className="inner">try it</div>
                 </button>
-                <img src={mjolnir} alt="mjolnir" className="randomchar__decoration" />
+                <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
             </div>
         </div>
-    );
+    )
 }
 
-// Компонент для отображения информации о персонаже
-const View = ({ char }) => {
-    const { name, description, thumbnail, homepage, wiki } = char;
-    let imgStyle = { 'objectFit': 'cover' };
+const View = ({data}) => {
+    const {name, description, thumbnail, homepage, wiki} = data;
+    let imgStyle = {'objectFit' : 'cover'};
     if (thumbnail === 'http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') {
-        imgStyle = { 'objectFit': 'contain' };
+        imgStyle = {'objectFit' : 'contain'};
     }
 
     return (
         <div className="randomchar__block">
-            <img src={thumbnail} alt="Random character" className="randomchar__img" style={imgStyle} />
+            <img src={thumbnail} alt="Random character" className="randomchar__img" style={imgStyle}/>
             <div className="randomchar__info">
                 <p className="randomchar__name">{name}</p>
                 <p className="randomchar__descr">
@@ -106,7 +77,7 @@ const View = ({ char }) => {
                 </div>
             </div>
         </div>
-    );
+    )
 }
 
 export default RandomChar;
